# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PRISM Writer API
  description: |
    RAG 기반 지능형 글쓰기 도구 백엔드 API
    
    ## 주요 기능
    - 목차 생성 (Outline Generator)
    - 참조 관리 (Reference Linking)
    - 문서 검색 (RAG Search)
    
  version: 0.1.0
  contact:
    name: PRISM Writer Team
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: 개발 서버
  - url: https://api.prism-writer.com
    description: 프로덕션 서버 (예정)

tags:
  - name: System
    description: 시스템 상태 확인
  - name: Outline
    description: 목차 생성 관련 API
  - name: References
    description: 참조 관리 관련 API

paths:
  # ============================================================================
  # System Endpoints
  # ============================================================================
  /health:
    get:
      tags:
        - System
      summary: 시스템 상태 확인
      description: 서비스 상태를 확인합니다.
      operationId: healthCheck
      responses:
        '200':
          description: 정상 동작
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: prism-writer-api
                  version:
                    type: string
                    example: 0.1.0

  # ============================================================================
  # Outline Endpoints
  # ============================================================================
  /v1/outline/generate:
    post:
      tags:
        - Outline
      summary: 목차 생성
      description: 주제와 참조 문서를 기반으로 AI가 목차를 생성합니다.
      operationId: generateOutline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutlineGenerateRequest'
      responses:
        '200':
          description: 목차 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutlineGenerateResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/outline/templates:
    get:
      tags:
        - Outline
      summary: 목차 템플릿 목록
      description: 사전 정의된 목차 템플릿 목록을 반환합니다.
      operationId: getOutlineTemplates
      responses:
        '200':
          description: 템플릿 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/OutlineTemplate'

  # ============================================================================
  # References Endpoints
  # ============================================================================
  /v1/drafts/{draft_id}/references:
    post:
      tags:
        - References
      summary: 참조 추가
      description: 글의 특정 문단에 청크 참조를 추가합니다.
      operationId: createReference
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
          description: 글 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceCreateRequest'
      responses:
        '201':
          description: 참조 추가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceResponse'
        '409':
          description: 중복 참조 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - References
      summary: 참조 목록 조회
      description: 글에 연결된 모든 참조 목록을 조회합니다.
      operationId: getReferences
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
          description: 글 ID
      responses:
        '200':
          description: 참조 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceWithContentResponse'

  /v1/drafts/{draft_id}/references/{reference_id}:
    delete:
      tags:
        - References
      summary: 참조 삭제
      description: 특정 참조를 삭제합니다.
      operationId: deleteReference
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
          description: 글 ID
        - name: reference_id
          in: path
          required: true
          schema:
            type: string
          description: 참조 ID
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 참조 없음

# ==============================================================================
# Components
# ==============================================================================
components:
  schemas:
    # --------------------------------------------------------------------------
    # Outline Schemas
    # --------------------------------------------------------------------------
    OutlineGenerateRequest:
      type: object
      required:
        - topic
      properties:
        topic:
          type: string
          minLength: 1
          maxLength: 500
          description: 글의 주제
          example: AI 시대의 글쓰기 방법론
        document_ids:
          type: array
          items:
            type: string
          description: 참조할 문서 ID 리스트
          default: []
        max_depth:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
          description: 목차 최대 깊이

    OutlineGenerateResponse:
      type: object
      properties:
        outline:
          type: array
          items:
            $ref: '#/components/schemas/OutlineItem'
        topic:
          type: string
        sources_used:
          type: integer

    OutlineItem:
      type: object
      properties:
        title:
          type: string
          description: 목차 제목
        depth:
          type: integer
          minimum: 1
          maximum: 5
          description: 목차 깊이 (1=H1, 2=H2, ...)

    OutlineTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        outline:
          type: array
          items:
            $ref: '#/components/schemas/OutlineItem'

    # --------------------------------------------------------------------------
    # Reference Schemas
    # --------------------------------------------------------------------------
    ReferenceCreateRequest:
      type: object
      required:
        - chunk_id
        - paragraph_index
      properties:
        chunk_id:
          type: string
          description: 참조할 청크 ID
        paragraph_index:
          type: integer
          minimum: 0
          description: 문단 인덱스 (0부터 시작)
        reference_type:
          type: string
          enum: [citation, summary, quote]
          default: citation
          description: 참조 유형

    ReferenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        draft_id:
          type: string
        chunk_id:
          type: string
        paragraph_index:
          type: integer
        reference_type:
          type: string
        created_at:
          type: string
          format: date-time

    ReferenceWithContentResponse:
      allOf:
        - $ref: '#/components/schemas/ReferenceResponse'
        - type: object
          properties:
            chunk_content:
              type: string
              description: 청크 텍스트 내용
            chunk_source:
              type: string
              description: 청크 출처 문서명

    # --------------------------------------------------------------------------
    # Common Schemas
    # --------------------------------------------------------------------------
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: 에러 메시지